# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Duet API
  version: 0.1.0
  description: |
    duetd 对外暴露的全部接口。Tauri 壳与浏览器走的是同一套 —— 壳不得通过 IPC 绕过 HTTP。

    **错误文案**：`Problem.type` 是机器可读错误码，前端据此查 i18n 词条；
    `detail` 只给开发者看，界面不展示。后端绝不返回用户可见的中文文案（见 docs/rules/i18n.md §3）。
servers:
  - url: http://127.0.0.1:{port}/v1
    variables:
      port:
        default: '7777'
security:
  - bearerAuth: []
...
paths:
...
  /system/update/prepare:
    post:
      tags: [system]
      operationId: prepareUpdate
      summary: 更新前暂停所有工作并保存检查点
      description: |
        ★ 这不是形式步骤，是一次跨聚合的领域操作。用户点「一键更新并重启」后
        **先调本接口，再下载**。

        对每个非终态的 Work：两段式取消 → 采集证据与事件游标 → 落检查点 → `paused`。

        返回 `blocked` 时前端**不得继续安装**，要把卡住的工作显示给用户。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpdatePrepareResult' }
        default: { $ref: '#/components/responses/Problem' }
...
  /events:
    get:
      tags: [system]
      operationId: streamEvents
      summary: 事件流（SSE）
      description: |
        ★ **这是 SSE，不是普通 JSON 响应**：`text/event-stream`，长连接。

        断线重连时带 `Last-Event-ID`（标准 SSE 头），服务端**只补它之后的**。
        从头补的话用户会看到整条时间线重放一遍；补少了则中间有洞，
        而洞是看不出来的——用户不知道自己漏了什么。

        每条 SSE 消息的 `id` 就是事件的 `seq`，浏览器的 EventSource 会自动
        把它记进 `Last-Event-ID`，重连时原样带回来。
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: integer, format: int64 }
          description: 上次收到的最后一个 seq；不带表示全新连接
      responses:
        '200':
          description: 事件流
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  每条形如：

                      id: 42
                      event: message_chunk
                      data: {"id":"evt_...","seq":42,...}

                  `data` 是一个完整的 Event 对象。
        default: { $ref: '#/components/responses/Problem' }

  # ── Project ─────────────────────────────────────────────────
  /projects:
    get:
      tags: [project]
      operationId: listProjects
      summary: 已添加的本地项目
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [project]
      operationId: addProject
      summary: 把一个本地代码文件夹加进来
      description: |
        ★★ **默认往用户的项目目录里写零个字节**，只登记路径。

        顺手初始化 `.acpflows/` 是很自然的想法，但用户刚把自己的仓库加进来、
        `git status` 就多出一堆没见过的东西，是最快失去信任的方式。

        ★ 要初始化就传 `initialize: true`，而**那之前必须先调
        `/projects/preview` 把要做的事讲给他听**。这条不是建议：
        「先说再做」是 `M3` 的全部意义——防的是**静默写**，
        不是「永远不写」。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: 本地绝对路径
                  examples: ['/Users/me/work/my-app']
                initialize:
                  type: boolean
                  default: false
                  description: |
                    是否照 `/projects/preview` 给出的计划创建 `.acpflows/`
                    并追加 `.gitignore`。**默认不做。**
      responses:
        '201':
          description: 已添加
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        default: { $ref: '#/components/responses/Problem' }
  /projects/preview:
    post:
      tags: [project]
      operationId: previewProject
      summary: 创建项目前的预演——**只看不动**
      description: |
        ★★ **一个字节都不写。** 用户交出来的是他自己的代码仓库，
        所以「先说再做」不是一个体贴的附加功能，而是这一步的全部意义。

        返回四块，正好对应设计稿弹层的四个区块：
        将创建什么 · 将追加什么 · 发现的已有 Skill · GitHub remote 与 `gh` 状态。

        ★ 「将创建」的条目与 `POST /projects` 带 `initialize` 时**实际执行的
        是同一份计划**——预演与执行各算一遍的话它们必然漂移，
        而漂移的方向永远是「预演里没说的那件事被做了」。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, examples: ['/Users/me/work/my-app'] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPreview' }
        default: { $ref: '#/components/responses/Problem' }
  /projects/{id}:
    delete:
      tags: [project]
      operationId: removeProject
      summary: 移除项目（只取消登记，**不删用户的文件**）
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '204': { description: 已移除 }
        default: { $ref: '#/components/responses/Problem' }

  # ── Work ────────────────────────────────────────────────────
  /works:
    get:
      tags: [work]
      operationId: listWorks
      summary: 全部工作
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [works]
                properties:
                  works:
                    type: array
                    items: { $ref: '#/components/schemas/Work' }
    post:
      tags: [work]
      operationId: startWork
      summary: 对一个项目提需求，开一个工作
      description: |
        切一个独立 worktree、建 ACP 会话、开始澄清需求。

        ★ worktree 建在 `~/.acpflows/worktrees`，**不在用户的项目目录里**
        （见 docs/plan/open-questions.md Q30）。用户把代码目录交给 Duet 时，
        并没有同意我们在他的仓库里造一堆分支和目录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, prompt]
              properties:
                project:
                  type: string
                  description: 项目的本地绝对路径
                prompt:
                  type: string
                  description: 用户的需求原话
                base_ref:
                  type: string
                  description: |
                    从哪儿开分支：分支名或 commit。留空时用仓库当前 HEAD。

                    ★ 让用户选基线的理由：他可能想从 `develop` 开工，
                    而当前分支上正躺着他没提交完的东西。
                  examples: [develop, 7c1de98]
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Work' }
        default: { $ref: '#/components/responses/Problem' }
  /works/prepare:
    post:
      tags: [work]
      operationId: prepareWork
      summary: 开工前的仓库状态——**只看不动**
      description: |
        ★★ **一个字节都不写。** 返回的是设计稿「新建工作」弹层要显示的东西：
        仓库脏不脏、有哪些分支可以当基线、当前 HEAD 是什么。

        ★ 未提交改动**分已跟踪与未跟踪两个数**：合成一条的话，
        「我只是新建了几个还没 add 的文件」和「我改了正在跟踪的代码」
        会长得一模一样，而对用户是两件完全不同的事。

        ★ 仓库处在 rebase / merge 中途时**报错拒绝开工**——
        那时切 worktree 会把用户正在解的冲突丢在那儿。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project: { type: string, description: 项目的本地绝对路径 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkPreparation' }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/worktree:
    get:
      tags: [work]
      operationId: getWorkWorktree
      summary: 一个工作的 git 现场——右栏「工作区」照它渲染
      description: |
        ★★ **只读。** 返回分支、领先几个 commit、未提交改动（逐个文件带增删）、
        以及**本次工作产生的** commit。

        ★ 基线之前的 commit 不算：那些是用户自己的历史，
        混进来会让他以为 Duet 改了他早先的提交。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorktreeState' }
        default: { $ref: '#/components/responses/Problem' }

  # ── Runtime ─────────────────────────────────────────────────
  /works/{id}/cancel:
    post:
      tags: [work]
      operationId: cancelWork
      summary: 停下这个工作正在跑的那一轮
      description: |
        ★ **不是所有状态都能停。** 审查中的工作会被拒绝（`work_cancel_not_allowed`）——
        独立审查正在产出证据，半路掐掉的话那个单元既没通过也没被驳回，
        卡在一个说不清的状态里。

        正常停下之后工作进入 `paused`，并落一条 `checkpoint` 事件——
        用户回头想接着干，得有东西告诉他「上次停在哪」。

        ★ **Agent 停不下来时会被强制结束**，工作进入 `failed`。
        这一步是「界面说已取消、后台还在烧钱改文件」的唯一防线：
        只报错不杀的话，用户以为什么都没发生。那时本端点仍返回错误，
        但工作状态与事件都已经落好。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 已停下
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/messages:
    post:
      tags: [work]
      operationId: sayInWork
      summary: 在这个工作里接着说一句
      description: |
        ★★ **同一个工作、同一条会话。** 这正是「连着说三句，AI 记得前两句」
        的那条路：会话按「工作 + 角色」常驻（Q42），第二句进的是上一句的上下文。

        ★ 不走这个端点而是再 `POST /works` 的话，用户说第二句时开的是一个
        **新工作**：新 worktree、新会话、新时间线——前一句彻底不在上下文里，
        而他以为自己只是补充了一句。

        用户那句话**原样**进时间线（一条 `user_message` 事件），
        然后后台跑一轮。返回 202 表示「收到了，正在跑」——
        一轮要好几分钟，同步等的话请求早超时了。

        终态的工作（`completed` / `failed` / `initializing_failed`）拒绝，
        返回 409 `work_not_accepting_messages`：那时再说什么都不会有人听，
        而界面上如果静默成功，用户会对着一个永远不动的时间线干等。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: 用户说的那句话，**原样**转发给 Agent 并进时间线
      responses:
        '202':
          description: 收到了，正在跑
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/requirement:
    get:
      tags: [work]
      operationId: getWorkRequirement
      summary: 这个工作当前的需求快照
      description: |
        ★ 还没有需求时返回 404 —— 那是新工作的常态，界面据此不显示标签，
        而不是显示一个「v0」。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        default: { $ref: '#/components/responses/Problem' }

    post:
      tags: [work]
      operationId: freezeWorkRequirement
      summary: 冻结当前这一版需求
      description: |
        ★★ **由用户点，不由 AI 判断**。AI 说「我觉得问清楚了」和用户说
        「就这样」是两件事——而冻结之后这一版就进了计划与契约，改不动了。

        ★ 还有待确认的事实时返回 409 `requirement_open_facts_remain`：
        带着没问清的问题往下走，AI 会自己替用户做决定，
        而那些决定会一路固化进计划与契约，等他发现时已经改了几十个文件。

        冻结之后再说一句话会产生 `v(n+1)` —— 冻结的那一版原样留着。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 已冻结
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/plan:
    get:
      tags: [work]
      operationId: getWorkPlan
      summary: 这个工作当前的计划（子计划 DAG + 单元）
      description: |
        ★ 还没规划时返回 404 —— 那是新工作的常态，界面据此不显示计划面板。

        ★★ 每个单元都带**由哪个角色做**（裁定三）。不带的话，
        用户看不出「这条谁在干」，而那正是他判断「该不该信这个产出」的依据
        （实现方审查自己的产出是 INV-ATT-8 明令禁止的）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        default: { $ref: '#/components/responses/Problem' }

    post:
      tags: [work]
      operationId: startWorkPlanning
      summary: 让计划架构师产出一版计划
      description: |
        ★★ **需求没冻结时拒绝**（INV-REQ-1），返回 409
        `requirement_not_frozen`：需求还在变的时候做出来的计划，做完也对不上，
        而那时用户已经等了一整轮还得从头再来。

        返回 202 表示「收到了，正在规划」——结果通过事件流回到界面。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: 收到了，正在规划
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/plan/history:
    get:
      tags: [work]
      operationId: getWorkPlanHistory
      summary: 计划的变更历史，**从新到旧**
      description: |
        ★ 设计稿计划面板的「变更历史」：改过几版、每次为什么改。
        旧版本一个字不改（INV-PLAN-4）——覆盖掉的话「上周那版拆成了什么」
        永远没有答案。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions:
                    type: array
                    items: { $ref: '#/components/schemas/Plan' }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/contract:
    get:
      tags: [work]
      operationId: getUnitContract
      summary: 一个单元的当前契约
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        default: { $ref: '#/components/responses/Problem' }

    post:
      tags: [work]
      operationId: designUnitContract
      summary: 让单元设计师产出一版契约
      description: |
        ★★ 契约由**单元设计师**产出，不是实现工程师自己写：自己给自己定
        验收标准与边界，等于没有边界——他会写一个刚好装得下自己想改的东西
        的范围（INV-ATT-8 的同一条道理）。

        产出的是**草稿**。冻结是用户的动作——自动冻结的话，用户还没看过
        这份契约，AI 就已经照着它开始改文件了。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: 收到了，正在设计 }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/contract/freeze:
    post:
      tags: [work]
      operationId: freezeUnitContract
      summary: 冻结这一版契约
      description: |
        ★ 一条验收标准都没有时拒绝冻结（409 `contract_empty`）：
        空契约冻结之后，「做完了」这件事没有任何判据——AI 说做完了就是做完了。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已冻结
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/acceptance:
    get:
      tags: [work]
      operationId: getUnitAcceptance
      summary: 一个单元的验收：标准与证据对照
      description: |
        ★★ 每条验收标准旁边是**它有没有证据**（设计稿的 `✓ ev-441` / `○ 无证据`）。

        没证据**不等于通过**——把它当成通过的话，一个什么都没做的单元
        也能「全部通过」。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acceptance' }
        default: { $ref: '#/components/responses/Problem' }

    post:
      tags: [work]
      operationId: collectUnitEvidence
      summary: 采集这个单元的 diff 证据
      description: |
        ★★ **应用自己去读 git，不问 AI**。让 AI 报告自己改了什么，等于让被
        考核的人填自己的考勤表——它不需要撒谎，只需要「记错了」一次，
        用户就再也不知道该信哪一条。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 采集完了
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acceptance' }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/accept:
    post:
      tags: [work]
      operationId: acceptUnit
      summary: 验收通过——提交改动并落检查点
      description: |
        ★★ **通过由用户点，不由 AI 判断。** 没有任何路径能自动走到这里。

        ★ 一条验收标准都没有证据时返回 409 `nothing_accepted`：
        允许的话，「验收」这个动作就没有内容了——用户点通过时以为自己
        核对过什么，而实际上什么都没有。

        ★ 没有改动时返回 409 `nothing_to_commit`，**不造空提交**：
        一个「验收通过」却什么都没改的单元，说明该被质疑的是那次验收。

        提交发生在**工作自己的 worktree** 上，用户的分支一字不动。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已提交
          content:
            application/json:
              schema:
                type: object
                required: [commit]
                properties:
                  commit:
                    type: string
                    description: 那次提交的短 sha，检查点绑着它
                    examples: ['abc1234']
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/start:
    post:
      tags: [work]
      operationId: startUnit
      summary: 让这个单元开工
      description: |
        ★★ 契约**没冻结时拒绝**（409 `contract_not_frozen`）：没冻结就开工
        的话，AI 干到一半契约变了，而它已经照着旧的那份改了十几个文件——
        产出对不上任何一版契约。

        这一轮由**单元自己派的那个角色**跑（裁定三）。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: 收到了，正在做 }
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/permission:
    post:
      tags: [work]
      operationId: answerPermission
      summary: 应答一次权限请求
      description: |
        Agent 想动用户的东西时会发一条 `request_permission` 事件到时间线，
        **并把这一轮挂起等应答**。用户在界面上点了某个选项，就走这个端点送回去。

        ★ `option_id` 是 **Agent 自己定义的不透明字符串**，从事件载荷的
        `options[].option_id` 原样取，**不要自己造、也不要按类别重新匹配**。
        搞错的话，用户点「拒绝」而 Agent 收到「允许」——这是整个权限体系里
        后果最严重的一种错。

        ★ 同一条请求**只能应答一次**。重复应答返回 409 且不再向 Agent 发第二条：
        Agent 那边只在等一个响应，多发的会被当成不认识的请求而静静丢弃，
        界面上却什么提示都没有。

        ★ 这个端点**不做超时**。用户可能去泡了杯咖啡；替他超时等于替他做决定。
        要中止只能靠取消整个工作（那时所有 pending 的请求都会收到 `cancelled`）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 工作标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ask_id, option_id]
              properties:
                ask_id:
                  type: string
                  description: 事件载荷里的 `ask_id`，标识是哪一条请求
                option_id:
                  type: string
                  description: |
                    用户选的那个选项的 `option_id`，**原样回传**。
      responses:
        '204':
          description: 已送达 Agent
        default: { $ref: '#/components/responses/Problem' }
...
  /skills:
    get:
      tags: [skill]
      operationId: listSkills
      summary: Skill 库
      parameters:
        - name: scope
          in: query
          required: false
          schema: { type: string, enum: [project, global] }
          description: 不传时返回全局库
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [skills]
                properties:
                  skills:
                    type: array
                    items: { $ref: '#/components/schemas/Skill' }
        default: { $ref: '#/components/responses/Problem' }
  /memories:
    get:
      tags: [memory]
      operationId: listMemories
      summary: 记忆库
      parameters:
        - name: scope
          in: query
          required: false
          schema: { type: string }
          description: |
            项目名（L2）或 `*`（L3 跨项目）。不传时返回全部。
            ★ 项目之间**永不串味**（INV-MEM-1）——两个项目的约定常常正好相反。
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/MemoryStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [memories]
                properties:
                  memories:
                    type: array
                    items: { $ref: '#/components/schemas/Memory' }
        default: { $ref: '#/components/responses/Problem' }
  /memories/{id}/review:
    post:
      tags: [memory]
      operationId: reviewMemory
      summary: 人工审核一条候选记忆
      description: |
        ★★ **绝不自动写入**（INV-MEM-2）。`candidate → active` 只能走这个端点，
        且必须带上 `actor`——AI 没有任何路径能自己把候选变成生效。

        错了的后果不是「多一条记忆」，而是 AI 把自己的一次臆断
        变成了以后每一轮的前提，而用户从没看过那句话。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, actor]
              properties:
                decision:
                  type: string
                  enum: [confirm, reject]
                actor:
                  type: string
                  description: 是谁做的决定。**必填**，空值一律拒绝
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Memory' }
        default: { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        duetd 启动时生成一次性 token，写入 `~/.acpflows/runtime/session.json`（0600）。
        无 token 一律 401 —— 防止本机其他程序静默驱动 Agent 写用户的代码。
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^proj-[0-9A-Za-z]+$'
      description: 项目标识，形如 `proj-01`
...
  responses:
    Problem:
      description: 错误（RFC 9457）
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    UpdatePrepareResult:
      type: object
      required: [status, prepared, blocked]
      properties:
        status:
          type: string
          enum: [ready, blocked]
          description: blocked 时前端不得继续安装
        prepared:
          type: array
          items:
            type: object
            required: [work_id, checkpoint_id]
            properties:
              work_id: { type: string, examples: ['work-08'] }
              checkpoint_id: { type: string, examples: ['ck-09'] }
        blocked:
          type: array
          items:
            type: object
            required: [work_id, reason]
            properties:
              work_id: { type: string }
              reason:
                type: string
                description: 机器可读原因码，前端翻译
                examples: [cancel_timeout, runtime_unresponsive]
              elapsed_seconds: { type: integer }
...
    Project:
      type: object
      required: [id, name, path, is_git_repo]
      properties:
        id: { type: string, examples: ['proj-01'] }
        name:
          type: string
          description: 目录名，展示用；用户可改而不影响 path
        path:
          type: string
          description: 本地绝对路径
        is_git_repo:
          type: boolean
          description: |
            ★ 不是 git 仓库**也允许添加**，只是标出来。
            直接拒绝的话，用户得先去命令行 `git init` 才能用这个产品——
            而他很可能正是不想碰命令行才来用的。
        default_branch:
          type: string
          description: 仓库的默认分支；非 git 仓库时缺省
        remedy:
          type: object
          description: 需要用户做点什么时给出的整条命令，与 Runtime.remedy 同一套做法
          properties:
            command:
              type: string
              examples: ['git init']

    # ── Work ──────────────────────────────────────────────────
    Work:
      type: object
      required: [id, state]
      description: |
        一次工作。`state` 是**封闭枚举**，与 backend/internal/constant 的
        WorkState 一一对应；新增取值要同时改三处：本文件 + 那个常量 +
        状态迁移表（domain/model/work.go）。
      properties:
        id: { type: string, examples: ['work-01'] }
        state:
          type: string
          description: |
            ★ `initializing_failed` 是**终态、不可恢复** —— worktree 没切成
            就没有可执行的现场（docs/adr/0006 Q1）。界面上不该给「重试」按钮。
          enum:
            - initializing
            - initializing_failed
            - clarifying
            - planning
            - ready
            - executing
            - reviewing_unit
            - waiting_user
            - paused
            - completed
            - failed
        project:
          type: string
          description: 项目的本地绝对路径
        worktree:
          type: string
          description: 这个工作的独立工作区；**在用户项目之外**
        prompt:
          type: string
          description: 用户的需求原话
        title:
          type: string
          description: |
            列表里显示的名字，**取自用户提的那句需求**（截断）。

            ★★ 左栏显示 `work-01` 的话，用户看不出那条工作是干嘛的——
            而他可能同时开着五六条。

            ★ 不用 AI 起的名字：起的名字与他说的话对不上时，
            他在列表里找不到自己那条工作。
          examples: ['用户能取消正在运行的 turn，取消后现场证据要保留']

    # ── Runtime ───────────────────────────────────────────────
    WorktreeState:
      type: object
      required: [branch, ahead, changes, commits]
      properties:
        branch: { type: string, examples: ['duet/work-08'] }
        base_commit:
          type: string
          description: |
            这个工作的起点。★ 空串表示**不知道基线**——
            那时 `ahead` 与 `commits` 都不可信，界面不该显示它们。
          examples: ['7c1de90']
        ahead:
          type: integer
          description: 相对基线领先几个 commit
        changes:
          type: array
          description: |
            未提交的改动，**逐个文件带增删行数**。
            只说「改了 3 个文件」的话，用户判断不出这次改动有多大。
          items: { $ref: '#/components/schemas/FileChange' }
        commits:
          type: array
          description: 本次工作产生的 commit，最近的排最前
          items: { $ref: '#/components/schemas/CommitInfo' }
...
    FileChange:
      type: object
      required: [path, added, removed]
      properties:
        path: { type: string }
        added: { type: integer }
        removed: { type: integer }
...
    WorkPreparation:
      type: object
      required: [current_branch, branches, head_commit, tracked_dirty, untracked]
      properties:
        current_branch: { type: string, examples: [main] }
        branches:
          type: array
          items: { type: string }
          description: 本地分支，可以当基线
        head_commit: { type: string, examples: ['7c1de98'] }
        tracked_dirty:
          type: integer
          description: |
            **已跟踪文件**里被改动的数量（含已暂存）。
            ★ 与未跟踪分开数——对用户是两件不同的事。
        untracked:
          type: integer
          description: 未跟踪文件的数量（逐个文件数，不是逐个目录）
    ProjectPreview:
      type: object
      required: [path, is_git_repo, actions, skills]
      properties:
        path: { type: string }
        name:
          type: string
          description: 默认显示名，取目录名
        is_git_repo:
          type: boolean
          description: |
            ★ 不是 git 仓库时**如实报告并继续**，绝不擅自 `git init`——
            在别人的目录里建仓库是不可逆的，而他可能有自己的打算。
        actions:
          type: array
          description: 将要做的每一步，**按执行顺序**
          items: { $ref: '#/components/schemas/ProjectAction' }
        skills:
          type: array
          description: |
            在项目里发现的已有 Skill（扫 `**/skills`，跳过 node_modules 与 target）。
            每条带**项目内的相对路径**当来源——用户要能照着去找。
          items: { $ref: '#/components/schemas/Skill' }
        remote: { $ref: '#/components/schemas/GitRemote' }
        gh: { $ref: '#/components/schemas/GhStatus' }
...
    MemoryStatus:
      type: string
      enum: [candidate, active, discarded, invalid, obsolete]
      description: |
        ★ **五态**。设计稿筛选器只有三档（active / 候选 / 已失效），
        那是**界面的分组**：「已失效」同时装着 `invalid` 与 `obsolete`——
        对用户长得一样，对系统不一样（废弃要带理由、可指向 supersedes）。
...